pyproject.toml
[tool.poetry]
name = "inverse-design"
version = "0.1.0"
description = ""
authors = ["pohaoc2 <pohaoc2@uw.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
matplotlib = "^3.10.0"
hydra-core = "^1.3.2"
black = "^24.10.0"
scipy = "^1.15.1"
pandas = "^2.2.3"
pylint = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
